+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Fixed:
 Changed:
 - Cleanup of unused docker images on GFZ servers is now done via `docker image prune`
 - Usage of the absolute url to checkout the git submodule for the CV
+- Improved style of items in device select of a data linking
 
 ## 1.16.1 - 2024-08-06